Clean up and fix SqlServerCe
